big changes!
okay, so full instillation has been put together and all assets are now in the level.  that said, no functionality has been set up yet.  the music plays on start (it's attached to the player pawn) and the camera sequence is still being set up (along with accompanying triggers).  the floor glow should probably go inside the player pawn BP and trigger with the audio and points on the floor (designated by boxes).  so basically still some work to do! :P